handle page refresh in admin pages
